Refactor: split test_weight_update.py into a helpers module + 4 files
The 890-line monolith had grown to 28 tests across model primitives, disk,
tensor/colocate, coordination, security, and CUDA-IPC. Split for readability +
maintainability, preserving every test verbatim (same logic/assertions):

- tests/_weight_update_helpers.py: shared factories/fixtures (tiny CPU model +
  continuous-serving stack, HTTP adapter, HF-named tensor sets with split q/k/v,
  flattened-bucket serialization, real-safetensors checkpoint writer). Underscore
  prefix so pytest does not collect it; imported after importorskip('torch').
- tests/test_weight_update_model.py: model-layer primitives + pure unit tests
  (update by name, atomic validate-then-copy, dry-run no-mutation, bucket
  round-trip, HF->logical name mapping). [5]
- tests/test_weight_update_disk.py: update_weights_from_disk executor/HTTP +
  get_weights_by_name + guards (swap/version/cache/abort, incompatible-config
  reject, no-engine guard, 200/400/403). [10]
- tests/test_weight_update_tensor.py: colocate path -- facade/worker delegation,
  HTTP route, slime chunked + empty-bucket partial semantics, in-flight guard,
  pickle-RCE allowlist (canary), malformed-400, real CUDA-IPC cross-process
  round-trip. [9]
- tests/test_weight_update_coordination.py: pause/continue/flush_cache/
  get_weight_version + the full slime HTTP flow + gating + flush-400. [4]

All 28 pass; broader serving suites green.
